requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
sqlalchemy==2.0.23
openai==1.3.5
python-dotenv==1.0.0
lxml==4.9.3
webdriver-manager==4.0.1

